minor modifications
git-svn-id: https://svn.kflog.org/svn/repos/cumulus/qt4/trunk@3034 c8f415cb-fa42-0410-b21e-89df95f64ae5

+++ cumulus/singlepoint.h
@@ -54,7 +54,7 @@ class SinglePoint : public BaseMapElement
               const WGSPoint& wgsPos,
               const QPoint& pos,
               const unsigned int elevation = 0,
-              const unsigned int secID=0 );
+              const unsigned short secID=0 );
 
   /**
    * Destructor
